requirements-base.txt
Django==1.11.16 # pyup: >=1.11,<2.0
Wand==0.4.4
amqp==1.4.9 # pyup: <2
beautifulsoup4==4.6.3
billiard==3.3.0.23 # pyup: < 3.4
bleach==2.1.4
celery==3.1.26.post2 # pyup: <4
git+https://github.com/mytardis/django-bootstrap-form.git@mytardis-dj1.9-bootstrap-v2#egg=django-bootstrap-form
django-extensions==1.8.1
git+https://github.com/mytardis/django-form-utils.git@django-1.11-upgrade#egg=django-form-utils
django-haystack==2.8.1
django-jstemplate==1.3.8
django-npm==1.0.0
django-registration-redux==2.4
# django-tastypie
git+https://github.com/mytardis/django-tastypie.git@mytardis-dj1.9#egg=django-tastypie
git+https://github.com/mytardis/django-tastypie-swagger.git@mytardis-dj1.9#egg=django-tastypie-swagger
django-user-agents==0.3.2
elasticsearch==1.9.0 # pyup: <2
gunicorn==19.7.1
gevent==1.3.6
html2text==2018.1.9
html5lib==1.0.1
isodate==0.6.0
pyjwt==1.6.4
kombu==3.0.37 # pyup: <4
lxml==4.2.5
paramiko==2.4.2
pillow==5.2.0
pwgen==0.8.2.post0
pyoai==2.5.0
pysolr==2.1.0 # pyup: >=2.1.0,<3
pystache==0.5.4
python-dateutil==2.7.3
# python-magic
git+https://github.com/mytardis/python-magic.git#egg=python-magic
pytz==2018.5
rdflib==4.2.2
requests==2.19.1
six==1.11.0
urllib3==1.23
wsgiref==0.1.2
python-ldap==2.4.27
pyyaml